Move implementation from Definition to Codeset
